Changing floats to integers

+++ main.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
-float subtract(float num1,float num2);
-float add(float num1,float num2);
+int subtract(float num1,float num2);
+int add(float num1,float num2);
 float multiplication(float num1, float num2);
 float division(float num1, float num2);
